+++ app/page.tsx
@@ -1,5 +1,5 @@
 import type React from "react"
-import { Carousel } from "./components/carousel"
+import { Carousel } from "./components/layout/carousel"
 
 export default function Home(): React.JSX.Element {
   return <Carousel />
